+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.9.1
+- `IFS-875`: [isy-aufrufkontext], [isy-serviceapi-core] Umsetzung des Transports von OAuth 2 Bearer-Token zwischen Schnittstellentechnologien
 - `IFS-1003`: [isy-sonderzeichen]
   + Refactoring des DIN-SPEC-Packages von isy-sonderzeichen
   + Entfernen des `core`-Unterpakets des DIN-SPEC-Packages von isy-sonderzeichen